[TEST] [#1049] Restructure Array tests - combine in one parent class
